backend/requirements_advanced.txt
# Dependencias para detección avanzada de manos con contexto de personas/caras
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0
flask
flask-sock
python-dotenv
# Librería adicional para mejor procesamiento de imágenes
scikit-image>=0.19.0
